Move Email.js to its new home
